Support lookup of string prop in AstNameTy
previously, lookup of string property in AstNameTy did not work and just
returned AstTopTy. For example, in es2022, there is a lookup of

  let body = script.ScriptBody

And according to the [spec](https://tc39.es/ecma262/#prod-ScriptBody),
this should evaluate to Ast[StatementList] but did not, causing imprecision
in the analysis.

This commit fixes the issue.